+++ src/Feature/Contacts/Groups/Bag/FindGroupsBag.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Smsapi\Client\Feature\Contacts\Groups\Bag;
+
+/**
+ * @api
+ */
+class FindGroupsBag
+{
+
+}